better database integrity support
- There is a problem when foreign keys are used and for example dropdown
is changed to empty value:

WP-MVC tries to write "" value which results into 0 written without
foreign keys used; this is acceptable but not quite correct.

When foreign key is used this try of course fails if you don't have
a 0 Id in the table; which is not quite common.

The fix just makes from "" value NULL value which makes update operation
more correct.