Added translation capability to wp-mvc and plugins based on it.
Created the directory core/languages and added translation for English and Spanish

Even though the first part of the link [0] tells you not to use a variable as
the text-domain, there are some parts where I believe this is a necessity
because the framework has to use the text-domain of the plugin not the
framework itself.

You'll find some parts where the text-domain is 'wpmvc'
And others is $this->plugin_name for example in the file mvc_helper.php
in the function admin_header_cells

This way you define the column translations in your own plugin's .po files, not in wpmvc

Only thing left to do in your plugin is to add the plugin_name object
to the registry, in the file bootstrap.php add something like the following:

  $mypname = 'my-app'; # this is the text domain
  MvcObjectRegistry::add_object('plugin_name', $mypname);

Then in the .po files use this file location for the list columns:

  #: helpers/mvc_helper.php:170
  msgid "Service Category Id"
  msgstr "Categoria de Servicio"

Use the returned value from wpmvc for the column as msgid (it makes everything easier)
in this case wpmvc generates the column name "Service Category Id"
for a table field called service_category_id
Then just translate "Service Category Id" to the desired value in your language.

In my tests I haven't had to define other locations "inside" wpmvc, for views
use the line in the view that calls the $this->form->input and it should be fine.

This approach doesn't work very well with automatic .po generation as the label is usually
not passed from your plugin and is created in wpmvc, but at least helps a lot to translate a plugin

Only thing left is to try with more than one wpmvc plugin installed

[0] https://developer.wordpress.org/plugins/internationalization/how-to-internationalize-your-plugin/