issue-183-multisite-fix
Change to make wp-mvc be friendly with multisite install as well as
single site setups.

The core is changed to use get_site_option, update_site_option opposed
to get_option and update_option which are not multisite friendly.

With the new functions in use, the events-calendar-example plugin was
updated in order to be an example on how to make a plugin that is both
multsite friendly and single site friendly.

The new plugin ends up having their own sets of database tables for
their blog.  ({prefix}_tablename, e.g wp_350_events)